Move app page and make room list endpoint
