+++ backend/src/rooms.rs
@@ -2,9 +2,11 @@ pub mod error;
 mod in_memory;
 pub mod manager;
 mod repo;
+pub mod routes;
 pub mod storage;
 
 pub use error::Error;
 pub use in_memory::InMemoryStorage;
 pub use manager::RoomManager;
 pub use repo::DatabaseStorage;
+pub use routes::router;
